add HW 32++
