--- src/main/java/com/burda/hw25/hw32/model/CommonResource.java
@@ -1,5 +1,0 @@
-package com.burda.hw25.hw32.model;
-
-public class CommonResource {
-    int x=0;
-}
