+++ src/main/java/com/burda/hw32/hippodrome/Hippodrome.java
@@ -1,4 +1,4 @@
-package com.burda.hw25.hw32.hippodrome;
+package com.burda.hw32.hippodrome;
 
 import java.time.LocalTime;
 import java.util.Random;
@@ -12,10 +12,8 @@ public class Hippodrome extends Thread {
         final String name = Thread.currentThread().getName();
         final int sleep = random.nextInt(400, 500);
         try {
-            for (int i = 0; i < 1000; i += 100) {
-                if (i == random.nextInt(100, 200)) {
-                    Thread.sleep(sleep);
-                }
+            for (int i = 0; i < 1000; i += random.nextInt(100, 200)) {
+                Thread.sleep(sleep);
             }
         } catch (InterruptedException e) {
             System.out.println(e.getMessage());
