+++ src/main/java/com/burda/hw32/hippodrome/Main.java
@@ -1,4 +1,4 @@
-package com.burda.hw25.hw32.hippodrome;
+package com.burda.hw32.hippodrome;
 
 public class Main {
     public static void main(String[] args) {
