+++ src/main/java/com/burda/hw32/hippodromeconcurrent/Hippodrome.java
@@ -1,4 +1,4 @@
-package com.burda.hw25.hw32.hippodromeconcurrent;
+package com.burda.hw32.hippodromeconcurrent;
 
 import java.util.Random;
 import java.util.concurrent.Callable;
@@ -11,10 +11,8 @@ public class Hippodrome implements Callable<String> {
         final String name = Thread.currentThread().getName();
         final int sleep = random.nextInt(400, 500);
         try {
-            for (int i = 0; i < 1000; i += 100) {
-                if (i == random.nextInt(100, 200)) {
-                    Thread.sleep(sleep);
-                }
+            for (int i = 0; i < 1000; i += random.nextInt(100, 200)) {
+                Thread.sleep(sleep);
             }
         } catch (InterruptedException e) {
             System.out.println(e.getMessage());
