+++ src/main/java/com/burda/hw32/model/CommonResource.java
@@ -1,0 +1,8 @@
+package com.burda.hw32.model;
+
+import lombok.Getter;
+
+@Getter
+public class CommonResource {
+    int x = 50;
+}
